removing the "simple" version of the model and created (ugly) Ui components for basic manipulations like : - creating employees - creating labels - editing employees - editing labels - editing employees planning - editing employees calendar (only for 2022 right now)

+++ src/main/java/fr/lesaffrefreres/rh/minibodet/model/DayLabelManager.java
@@ -14,9 +14,9 @@ public interface DayLabelManager {
 
     public void removeDayLabelById(long id);
 
-    public void setDayLabelText(int id, String txt);
+    public void setDayLabelText(long id, String txt);
 
-    public void setDayLabelColor(int id, Color c);
+    public void setDayLabelColor(long id, Color c);
 
     public boolean labelExist(long id);
